Update include/C-Utils/cutils.c in develop branch (2026-03-13).
Modified:
	include/C-Utils/cutils.c;

+++ include/C-Utils/cutils.c
@@ -73,7 +73,7 @@ int clear_stdin(void)
 
 	while(characters != 10 && characters != EOF)
 	{
-		characters = getchar();
+		characters = (short int)getchar();
 	}
 
 	return C_UTILS_SUCCESS;
